requirements.txt
# Core ML and geospatial dependencies
geoai-py==0.11.0
torch==2.8.0
torchvision==0.23.0
numpy==2.2.6
pandas==2.3.2
rasterio==1.4.3
rioxarray==0.19.0
geopandas==1.0.1
shapely==2.1.2
pyproj==3.7.2
fiona==1.10.1
xarray==2025.9.0
dask==2025.9.0
pystac==1.10.1
pystac-client==0.7.6
stackstac==0.5.1
scikit-image==0.25.2
albumentations==2.0.8
opencv-python-headless==4.12.0.88
requests==2.32.5
tqdm==4.67.1

# Optional extras live in requirements-dev.txt
